refactor: move discover and path helpers to utils
