+++ src/loaders/index.ts
@@ -2,4 +2,6 @@ import json from './json'
 import matter from './matter'
 import yaml from './yaml'
 
+export { defineLoader } from './types'
+
 export const loaders = [json, yaml, matter]
